# docs/swagger.yaml
basePath: /
definitions:
  github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse:
    description: Standard API error response
    properties:
      errors:
        additionalProperties:
          type: string
        description: Detailed field errors
        type: object
      message:
        description: Human readable summary
        type: string
      status:
        description: HTTP Status Code
        type: integer
    type: object
  internal_resources_deployment.Deployment:
    description: Deployment entity representing a built artifact with content-addressable
      identifier
    properties:
      created_at:
        type: string
      expired_at:
        description: 'Nullable: some builds may never expire'
        type: string
      hash:
        description: The content-addressable identifier
        type: string
      id:
        type: string
      project_id:
        type: string
      status:
        type: string
    type: object
  internal_resources_deployment.DeploymentPaged:
    description: Paginated deployment response with metadata
    properties:
      items:
        items:
          $ref: '#/definitions/internal_resources_deployment.Deployment'
        type: array
      pageCount:
        type: integer
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
    type: object
  internal_resources_project.CreateProject:
    description: Project creation DTO
    properties:
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - name
    type: object
  internal_resources_project.Project:
    description: Project entity representing a project with its metadata
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  internal_resources_project.ProjectPaged:
    description: Offset-based paginated project response with metadata
    properties:
      items:
        items:
          $ref: '#/definitions/internal_resources_project.Project'
        type: array
      pageCount:
        type: integer
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
    type: object
  internal_resources_project.UpdateProject:
    description: Project update DTO
    properties:
      id:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - id
    type: object
host: localhost:8080
info:
  contact: {}
  description: The core API for the Infario Cloud Platform.
  title: Infario API
  version: "1.0"
paths:
  /deployments:
    get:
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: pageNumber
        type: integer
      - default: 25
        description: 'Page size (default: 25, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_resources_deployment.DeploymentPaged'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: List all deployments
      tags:
      - deployments
  /deployments/{id}:
    get:
      parameters:
      - description: Deployment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_resources_deployment.Deployment'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: Get a deployment by ID
      tags:
      - deployments
  /deployments/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Project ID
        in: formData
        name: project_id
        required: true
        type: string
      - description: Content-addressable hash
        in: formData
        name: hash
        required: true
        type: string
      - description: Binary file (zip or tar.gz)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_resources_deployment.Deployment'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: Upload a deployment artifact
      tags:
      - deployments
  /projects:
    get:
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: pageNumber
        type: integer
      - default: 10
        description: 'Page size (default: 10, max: 100)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_resources_project.ProjectPaged'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_resources_project.CreateProject'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_resources_project.Project'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: Create a project
      tags:
      - projects
  /projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: Delete a project
      tags:
      - projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_resources_project.Project'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: Get a project by ID
      tags:
      - projects
    patch:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Project Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_resources_project.UpdateProject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_resources_project.Project'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_dimasbaguspm_infario_pkgs_response.ErrorResponse'
      summary: Update a project
      tags:
      - projects
schemes:
- http
- https
swagger: "2.0"